Reorganized with configs
